✨ Add available-capacity.js and form-validation-confirm-modal.js
